+++ tests/contrib/__init__.py
@@ -1,7 +1,2 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, print_function, unicode_literals
-
-from importanize.plugins import deactivate_all_plugins
-
-
-deactivate_all_plugins()