+++ tests/test_data/input_unused_groups.py
@@ -1,0 +1,10 @@
+from __future__ import absolute_import
+import os
+
+import click
+
+from .foo import bar
+
+
+os
+bar