+++ tests/test_data/output_unused_groups.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import, print_function, unicode_literals
+import os
+
+from .foo import bar
+
+
+os
+bar